# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MyReadings API
  description: >
    API REST de MyReadings: gestión de lecturas familiares con sistema de
    recompensas. Roles PARENT (padre/madre) y CHILD (hijo/a). Todos los
    endpoints salvo `/auth/register` y `/auth/login` requieren JWT
    (`Authorization: Bearer <token>`).
  version: "1.0.0"
  contact:
    name: MyReadings (TFM)
servers:
  - url: http://localhost:3000/api
    description: Desarrollo local
  - url: /api
    description: Relativo (mismo host que el frontend)
...
security:
  - bearerAuth: []
paths:
...
  /books/{id}/request-reward:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    post:
      tags: [books]
      summary: Solicitar recompensa sobre un libro (hijo)
      description: >
        Solo CHILD, sobre libro propio en estado `NOT_STARTED`. Crea un
        `RewardRequest` (PENDING) y una `Notification` `REWARD_REQUEST`
        al padre.
      responses:
        "201":
          description: Solicitud creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RewardRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─────────────────────────── GOALS ───────────────────────────
...
  /goals/{id}/redeem:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    post:
      tags: [goals]
      summary: Canjear una meta alcanzada
      description: Solo PARENT; la meta debe estar en estado `ACHIEVED`.
      responses:
        "200":
          description: Meta canjeada (REDEEMED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  # ─────────────────────────── REWARDS ───────────────────────────
  /books/{bookId}/rewards:
    parameters:
      - name: bookId
        in: path
        required: true
        schema:
          type: integer
    post:
      tags: [rewards]
      summary: Crear una recompensa sobre un libro
      description: >
        Solo PARENT; libro `NOT_STARTED`; hijo propio; `deadline` futura;
        `goalId` obligatorio si `type = POINTS`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRewardRequest"
      responses:
        "201":
          description: Recompensa creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reward"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /rewards:
    get:
      tags: [rewards]
      summary: Listar recompensas
      description: Padre ve las de sus hijos; hijo ve las suyas.
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/RewardStatus"
      responses:
        "200":
          description: Lista de recompensas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reward"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /rewards/{id}:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    get:
      tags: [rewards]
      summary: Detalle de una recompensa
      responses:
        "200":
          description: Recompensa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reward"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [rewards]
      summary: Editar una recompensa
      description: Solo PARENT; el libro debe seguir `NOT_STARTED`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRewardRequest"
      responses:
        "200":
          description: Recompensa actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reward"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [rewards]
      summary: Eliminar una recompensa
      description: Solo PARENT; el libro debe seguir `NOT_STARTED`.
      responses:
        "204":
          description: Eliminada
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─────────────────────────── REWARD REQUESTS ───────────────────────────
  /reward-requests:
    get:
      tags: [reward-requests]
      summary: Listar solicitudes pendientes de mis hijos
      description: Solo PARENT.
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/RewardRequestStatus"
      responses:
        "200":
          description: Lista de solicitudes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RewardRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /reward-requests/{id}:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    patch:
      tags: [reward-requests]
      summary: Resolver o descartar una solicitud
      description: Solo PARENT; debe ser de un hijo suyo.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveRewardRequestRequest"
      responses:
        "200":
          description: Solicitud actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RewardRequest"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─────────────────────────── NOTIFICATIONS ───────────────────────────
  /notifications:
    get:
      tags: [notifications]
      summary: Listar notificaciones propias
      parameters:
        - name: read
          in: query
          schema:
            type: boolean
          description: Filtrar por leídas/no leídas
      responses:
        "200":
          description: Lista de notificaciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /notifications/unread-count:
    get:
      tags: [notifications]
      summary: Número de notificaciones no leídas
      responses:
        "200":
          description: Contador
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 3
        "401":
          $ref: "#/components/responses/Unauthorized"
  /notifications/{id}/read:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    patch:
      tags: [notifications]
      summary: Marcar una notificación como leída
      responses:
        "200":
          description: Notificación actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /notifications/read-all:
    patch:
      tags: [notifications]
      summary: Marcar todas las notificaciones propias como leídas
      responses:
        "200":
          description: Notificaciones actualizadas
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: integer
                    example: 5
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ─────────────────────────── LEDGER ───────────────────────────
  /ledger/balance:
    get:
      tags: [ledger]
      summary: Saldo propio (puntos y euros)
      description: Solo CHILD.
      responses:
        "200":
          description: Saldo actual
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Balance"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /children/{childId}/ledger:
    parameters:
      - name: childId
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [ledger]
      summary: Movimientos y saldo de un hijo
      description: Solo PARENT; el hijo debe ser suyo.
      responses:
        "200":
          description: Ledger del hijo
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    $ref: "#/components/schemas/Balance"
                  entries:
                    type: array
                    items:
                      $ref: "#/components/schemas/LedgerEntry"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─────────────────────────── STATS ───────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
      example: 1
  responses:
    BadRequest:
      description: Validación de datos fallida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Token ausente o inválido
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Rol o propiedad insuficiente
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflicto (email duplicado, meta ya canjeada, categoría con libros, etc.)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          example: "El color debe ser un hex #RRGGBB"
        error:
          type: string
          example: "Bad Request"
...
    GoalStatus:
      type: string
      enum: [ACTIVE, ACHIEVED, REDEEMED]
    RewardType:
      type: string
      enum: [POINTS, MONEY]
    RewardStatus:
      type: string
      enum: [PENDING, FULFILLED, PENALIZED]
    RewardRequestStatus:
      type: string
      enum: [PENDING, RESOLVED, DISMISSED]
    NotificationType:
      type: string
      enum: [REWARD_REQUEST, REWARD_FULFILLED, REWARD_PENALIZED, GOAL_ACHIEVED]
    LedgerKind:
      type: string
      enum: [POINTS, MONEY]
    LedgerReason:
      type: string
      enum: [FULFILLED, PENALTY, REDEMPTION]

    # ---------- Auth / Users ----------
...
    Goal:
      type: object
      properties:
        id:
          type: integer
          example: 3
        childId:
          type: integer
          example: 2
        createdByParentId:
          type: integer
          example: 1
        name:
          type: string
          example: "PlayStation"
        description:
          type: string
          nullable: true
        targetPoints:
          type: integer
          minimum: 1
          example: 1000
        status:
          $ref: "#/components/schemas/GoalStatus"
        createdAt:
          type: string
          format: date-time
      required: [id, childId, createdByParentId, name, targetPoints, status, createdAt]
...
    Reward:
      type: object
      properties:
        id:
          type: integer
          example: 7
        bookId:
          type: integer
          example: 10
        createdByParentId:
          type: integer
          example: 1
        type:
          $ref: "#/components/schemas/RewardType"
        value:
          type: number
          format: double
          example: 50
        deadline:
          type: string
          format: date
        penaltyValue:
          type: number
          format: double
          nullable: true
        goalId:
          type: integer
          nullable: true
        status:
          $ref: "#/components/schemas/RewardStatus"
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, bookId, createdByParentId, type, value, deadline, status, createdAt]
    CreateRewardRequest:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/RewardType"
        value:
          type: number
          format: double
          exclusiveMinimum: true
          minimum: 0
          example: 50
        deadline:
          type: string
          format: date
          description: Debe ser una fecha futura
        penaltyValue:
          type: number
          format: double
          nullable: true
        goalId:
          type: integer
          nullable: true
          description: Obligatorio si `type = POINTS`
      required: [type, value, deadline]
    UpdateRewardRequest:
      type: object
      properties:
        value:
          type: number
          format: double
          exclusiveMinimum: true
          minimum: 0
        deadline:
          type: string
          format: date
        penaltyValue:
          type: number
          format: double
          nullable: true
        goalId:
          type: integer
          nullable: true

    # ---------- RewardRequest ----------
    RewardRequest:
      type: object
      properties:
        id:
          type: integer
          example: 4
        bookId:
          type: integer
          example: 10
        childId:
          type: integer
          example: 2
        status:
          $ref: "#/components/schemas/RewardRequestStatus"
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, bookId, childId, status, createdAt]
    ResolveRewardRequestRequest:
      type: object
      properties:
        status:
          type: string
          enum: [RESOLVED, DISMISSED]
      required: [status]

    # ---------- Notifications ----------
    Notification:
      type: object
      properties:
        id:
          type: integer
          example: 12
        recipientUserId:
          type: integer
          example: 1
        type:
          $ref: "#/components/schemas/NotificationType"
        refBookId:
          type: integer
          nullable: true
        refRewardId:
          type: integer
          nullable: true
        message:
          type: string
          example: "recompensa pendiente de activar"
        read:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
      required: [id, recipientUserId, type, message, read, createdAt]

    # ---------- Ledger ----------
    LedgerEntry:
      type: object
      properties:
        id:
          type: integer
          example: 20
        childId:
          type: integer
          example: 2
        rewardId:
          type: integer
          nullable: true
        kind:
          $ref: "#/components/schemas/LedgerKind"
        amount:
          type: number
          format: double
          example: 50
        goalId:
          type: integer
          nullable: true
        reason:
          $ref: "#/components/schemas/LedgerReason"
        createdAt:
          type: string
          format: date-time
      required: [id, childId, kind, amount, reason, createdAt]
    Balance:
      type: object
      properties:
        points:
          type: integer
          example: 450
        money:
          type: number
          format: double
          example: 12.5
      required: [points, money]

    # ---------- Stats ----------
...